requirements.txt
Django==3.2.5
djangorestframework>=3.12,<4.0
mysqlclient==2.0.3